CellarItems now has 3 items and extensions
Items now have extensions. They're now in the format:
NAME, SPRITE, COLOUR, TYPE, EFFECT
Or, for armour:
NAME, SPRITE, COLOUR, TYPE, BODY PART, LEVEL
This allows a bonus to be applied across a type of armour, and it to be adjusted based on body part, as well as allowing for the equipping system to be a bit easier.

Also i'm using the number 6 for potions bc it looks like a splash potion from minecraft. Keeping scrolls as ?, going to use ^ for helmets, ] for chestplates, ( for leggings and [ for boots.

Colour indicates material or effect. For example, a water bottle is blue, a blank scroll is an off-white, and leather armour is brown.

That's all for now! Just a wee bit for the future of additions and all that!