Update CellarItems.csv
Never mind.
Potions now have two properties as well: an effect and a property when poured onto the ground (e.g. it smokes, it fizzes, it evaporates quickly, it changes colour) which is just for decorative purposes, unless maybe a potion gets poured onto some kind of object in the world the effect says something about whether or not the item is beneficial/harmful to drink?